moves fresh name generation from semantic to baseline
