+++ scalameta/semantic/src/main/scala/scala/meta/semantic/Api.scala
@@ -648,31 +648,6 @@ private[meta] trait Api {
     def isReference: Boolean = !isBinder
   }
 
-  private object Fresh {
-    private var id = new java.util.concurrent.atomic.AtomicInteger()
-    def nextId() = id.incrementAndGet()
-  }
-
-  implicit class XtensionSemanticTermModule(tree: Term.type) {
-    def fresh(): Term.Name = fresh("fresh")
-    def fresh(prefix: String): Term.Name = impl.Term.Name(prefix + Fresh.nextId())
-  }
-
-  implicit class XtensionSemanticTypeModule(tree: Type.type) {
-    def fresh(): Type.Name = fresh("fresh")
-    def fresh(prefix: String): Type.Name = impl.Type.Name(prefix + Fresh.nextId())
-  }
-
-  implicit class XtensionSemanticPatModule(tree: Pat.type) {
-    def fresh(): Pat.Var.Term = impl.Pat.Var.Term(Term.fresh().require[impl.Term.Name])
-    def fresh(prefix: String): Pat.Var.Term = impl.Pat.Var.Term(Term.fresh(prefix).require[impl.Term.Name])
-  }
-
-  implicit class XtensionSemanticPatTypeModule(tree: Pat.Type.type) {
-    def fresh(): Pat.Var.Type = impl.Pat.Var.Type(Type.fresh().require[impl.Type.Name])
-    def fresh(prefix: String): Pat.Var.Type = impl.Pat.Var.Type(Type.fresh(prefix).require[impl.Type.Name])
-  }
-
   // ===========================
   // PART 9: REPRESENTATION CONVERSIONS
   // ===========================
