+++ scalameta/trees/src/main/scala/scala/meta/internal/ast/Fresh.scala
@@ -1,0 +1,8 @@
+package scala.meta
+package internal
+package ast
+
+object Fresh {
+  private var id = new java.util.concurrent.atomic.AtomicInteger()
+  def nextId() = id.incrementAndGet()
+}
